[CDB][0743] Marked NPCs with already correct stats as ready to use new class level stats system
* Marked NPC with already correct level, health, mana and multipliers as
ready to use new class level stats system

With this update, Classic DB development is entering a transitional
phase where some creatures will use the new stats system while the others
still use current DB values. Those DB values can be correct, slightly
incorrect or wrong. We are in the process of checking _every_ NPC in the
game to fix their stats and move everyone of them to the new system.
This is a long process that will take weeks or months.

Some NPC with stats higher that they should will become easier, some others will
become tougher.
The overall game balance will be unchanged.

During this process, we may need to _temporary_ change some NPC stats to values
they had in TBC or later, but not in Classic.
We are aware of this possibility and those NPCs will be fixed soon after.

We want to stress out that by this process we are not changing the creatures stats.
We are fixing them. This will result in a clean and proper creatures database.

CP from https://github.com/cmangos/classic-db/commit/85330361101202b04c944701f3f553ca045246bf