[ACID] Adjustment to Crystalcore Mechanic AI
The proper behavior for Crystalcore Mechanic (20052) is to cast Recharge (37121) on Crystalcore Devastator (20040) or Crystalcore Sentinel (20041) when they cast Power Down (34937) at low health.
This is not really possible to do correctly with ACID currently, so this commit is a temporary solution.

EVENT_T_FRIENDLY_IS_CC partially works since Power Down stuns the caster, but the problem is the event will be triggered for other unrelated CC effects and creatures (not just Powered Down robots).

I also attempted to make the robots use ACTION_T_THROW_AI_EVENT at the same time they Power Down and then added EVENT_T_RECEIVE_AI_EVENT on the Mechanics to make them respond by casting Recharge on the event sender.
This approach is flawed as well because it's possible for out of combat Crystalcore Mechanics within the radius to receive the thrown AI event and get pulled/aggroed when they shouldn't.