Add missing dbscripts_on_quest_* to quest_template
Several quests had scripts in dbscripts_on_quest_* but was not set in quest_template, causing the script to not run.

We probably want to throw a DBError when this is the case. Currently all three DBs had this issue, but we don't know because there are no errors about it in the logfile.

@cala, @Grz3s I suggest running the below queries since there are some missing ones in classic and wotlk as well

SELECT * FROM dbscripts_on_quest_end WHERE id NOT IN (SELECT CompleteScript FROM quest_template);
SELECT * FROM dbscripts_on_quest_start WHERE id NOT IN (SELECT StartScript FROM quest_template);