Implement Hero of the Mag'har q.10212 CompleteScript
This is one of the longest events in WoW. It is based on this cinematic from WC3: https://www.youtube.com/watch?v=luuUE0x7-TY

It starts after a player completes Hero of the Mag'har q.10212 (changed to There Is No Hope q.10172 in Cataclysm). The event will only run once every 6 hours.

* Thrall, Drek'thar and an army of wolf riders spawn nearby Nesingwary Safari
* They ride all the way to Garadar Gates where the riders form up outside and Thrall and Drek'thar continue inside Garadar.
* Thrall goes to speak with Garrosh and Grandmother Geyah
* Thrall and Garrosh go to the blue summoning ring where the main event takes part
* After the event is finished Garrosh (or rather his buff bot triggers) buff all players and NPCs in Garadar with Hellscreams Warsong
* After this Garrosh goes back to his spawn position and Thrall goes to Grandmother Geyah and sits there for 6 hours before despawning

This closes: #63