Update damage for Prince Malchezaar's Axes (17650)
TBC-DB had 0.006 as the multiplier, causing the axes to only deal 1-2 damage per hit.
Since we can't find any evidence to support this multiplier, we'll use WotLK-DB data instead.
There are also alot of comments/guides which say the axes should have some kind of stacking buff which ramps up their damage over time, but we can't find this spell in sniffs if it exists.