Implementation of q.10985 A Distraction for Akama
This took awhile, so hopefully it works well. This not only contains the
scripting needed for the quest, it contains scripting for the RP battle
outside of BT.

I also added some missing NPCs and linked the Illidari forces to
Doomwalker, as they should not respawn if he's alive.

I am sure this implementation is not 100% correct. Though a lot of this
was looked at with sniff, a lot was also done by observing old youtube
videos and how the event happens on the retail PTR and guesstimating.

From all my testing I didn't see the quest really bug out at all. Shouldn't
crash, and the only reason it would get "stuck" is because Akama failed to
resume waypointing, but hasn't happened to me.