CDB-2671 Introduce Classic DB custom content
* Introduce support for SQL updates to DB contents that are known to be different from official/sniff data
These updates are applied to ensure expected functionning from a gameplay point of view
(aka "Blizzlike") when using CMaNGOS core.

These updates SHOULD NOT be part of a full DB dump (like for DB release) or inserted in a regular
SQL update file as we need to keep DB data as "pure" as possible to analyse them and
deduce proper core behaviour from them.

Possible but non-exhaustive reasons for these updates to be needed:
- CMaNGOS core does not properly handle the data these updates are fixing (yet)
- Official/sniff data are from a different expansion/patch and where possibly changed since Classic
- Official/sniff data from client are different from what they are server-side
- Data or their parsing were wrong

(cherry-picked from commit
https://github.com/cmangos/classic-db/commit/595590c553bfecbc502303129347932669efe47a)