Fix quest chaining for Children's week (Classic and TBC)
There has been previous iterations to these quests that left it in an inconsistent and incorrect state, so redo it completely instead.
All 4 chains have the same structure (1 leads to 3, leads to 2, leads to 1)

Also update MinLevel and QuestLevel. As far as I can tell the QuestLevel was not changed to -1 until WoTLK. Prior to that it was 60 for classic quests and 70 for TBC.

Please note that the TBC quests are still auto-completed and needs to be scripted