Make dbscripts_on_creature_movement.id = 2 truly GENERIC
Befor the Npc would return to their spawnposition and start to wander_distance there, all current usecases apply to the changed logic though

-- 19162, 13 last point, aligns with this change (-1674.3, 5287.94, -50.6398)-> (-1668.49, 5284.1, -49.4239)
-- 19155, 24 last point, aligns with this change (-1937.06, 5173.81, -40.2092)-> (-1939.99, 5179.95, -40.2092)
-- 20195, 16 last point, aligns with this change (430.253, 2896.79, 52.6058)-> (430.754, 2896.12, 52.6334)

We should think about adding wander_distance (automatically sets movementtype to 1 at a waypoint for the duration of waittime if set) to waypoints table.

we gutted all the columns out of these tables, leaving only xyzo waittime and script, but we should think about readding movement related columns back to it for sanity and reasons.

https://github.com/vmangos/core/commit/a2ba96ecf0cb77d4d19dd13fa359b8045184bd88