TBC PTR Vendors
Flaskataur and The Rulkster were used at different points during TBC to facilitate raid testing. We don't have the full list of items they offered, but we know both of them were vendors.
For Flaskataur, I found a screenshot of the first page of vendor items from when he was located in SWP.