Fix missing object needed for quest 8924 'Hunting for Ectoplasm'.
This is basically something that was forgotten when backporting commit
cmangos/wotlk-db@8c5dbb654fbb. This object was NOT added in that commit
though, because as mentioned in the associated pull request (see
cmangos/wotlk-db#594) both objects were already in the database, and
'the person who will backport it will simply take it from there'.
But that's very easy to miss of course, so it didn't happen...

This finally fixes cmangos/issues#3392 for real.